+++ source/Application/Controller/SuggestController.php
@@ -14,8 +14,7 @@ use OxidEsales\Eshop\Core\Registry;
  * Article suggestion page.
  * Collects some article base information, sets default recommendation text,
  * sends suggestion mail to user.
- *
- * @deprecated since v6.2.0 (2017-02-15); Recommendations feature will be moved to an own module.
+ * @deprecated since v6.5.4 (2020-04-06); Suggest feature will be removed completely
  */
 class SuggestController extends \OxidEsales\Eshop\Application\Controller\FrontendController
 {